# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IAM Integration Platform API
  version: "1.0.0"
  description: |
    API for the HKUST IAM Integration Platform POC.
    Supports ingestion of identity events from CADS, PeopleSoft (SIS/FMS/HRMS),
    ECM, and JSPM; event status tracking; user sync status queries; and
    real-time access decisions.

    **Authentication:** All endpoints require an `api_key` header.
    The default POC key is `poc-dev-key-1234`.

    **Email domains:** Currently only `@ust.hk` is accepted.
    Future expansion: `@connect.ust.hk`, `@family.ust.hk`.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/v1/inbound/events:
    post:
      summary: Ingest a single identity event
      description: |
        Unified ingestion endpoint for CADS, PeopleSoft, ECM, and JSPM.
        The source system is identified via the `X-Source-System` header or
        the `sourceSystem` field in the request body.

        Returns 202 Accepted with an `eventId` for status tracking.
        Returns 200 if the `Idempotency-Key` was already seen (duplicate).
      operationId: ingestEvent
      tags: [Ingestion]
      parameters:
        - $ref: '#/components/parameters/ApiKey'
        - $ref: '#/components/parameters/XSourceSystem'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestPayload'
            examples:
              cads:
                summary: CADS HR event
                value:
                  employeeId: "E12345"
                  employeeEmail: "john.doe@ust.hk"
                  employeeName: "John Doe"
                  department: "Finance Management Office"
                  role: "APPROVER"
                  action: "provision"
              peoplesoft_hrms:
                summary: PeopleSoft table-row event
                value:
                  "Dept": "DAO"
                  "Rank/ Team": "Alumni Team"
                  "User": "dao.alumni.manager"
                  "Role Name": "HKUST ALUM ADMIN DOWNLOAD DATA"
                  "Remarks": "Access to AAS"
                  "Data Level Security": "All alumni"
              ecm:
                summary: ECM access event
                value:
                  userId: "ECM-001"
                  userEmail: "alice.chan@ust.hk"
                  documentClass: "FINANCE_CONTRACTS"
                  role: "READER"
                  action: "provision"
              jspm:
                summary: JSPM project assignment
                value:
                  projectCode: "PROJ-2026-001"
                  userEmail: "bob.lee@ust.hk"
                  projectRole: "PROJECT_MANAGER"
                  action: "assign"
      responses:
        "202":
          description: Event accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        "200":
          description: Duplicate event (idempotency key already seen)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
...
  /api/v1/inbound/events/{eventId}/status:
    get:
      summary: Get event processing status
      description: |
        Returns the current processing lifecycle of a specific event.
        Status values: `received` → `validated` → `sent_to_midpoint` → `success`
        or `failed` → `retrying` → `dead_letter`
      operationId: getEventStatus
      tags: [Ingestion]
      parameters:
        - $ref: '#/components/parameters/ApiKey'
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventStatus'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

  # ── Source-specific aliases (same handler as /events) ──────────────────────
  /api/v1/inbound/cads:
    post:
      summary: Ingest CADS event (alias)
      description: |
        Alias for `POST /api/v1/inbound/events` with `X-Source-System: CADS`.

        Accepts **two input formats**:
        1. **Raw CADS table row** — object whose keys match CADS spreadsheet column headers
           (e.g. `"User Email"`, `"Role"`, `"Valid From"`).  The backend automatically detects
           this format and runs it through the CADS transformer before ingestion.
        2. **Canonical JSON** — object with top-level `meta`, `identity`, and `entitlement`
           keys (as returned by `POST /api/v1/inbound/cads/transform`).

        Returns 422 if a raw row fails required-field validation.
      operationId: ingestCads
      tags: [Ingestion]
      parameters:
        - $ref: '#/components/parameters/ApiKey'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestPayload'
            examples:
              cads_raw_row:
                summary: Raw CADS table row
                value:
                  "User Email": "user_a"
                  "Role": "BCO"
                  "Department / Project": "16500"
                  "(1)\nEnquire REQ/PO/\nReceipt ": "Y"
                  "(2)\nRecord Receipt of Goods/\nServices": "Y"
                  "(3)\nCertify Receipt for Payment": "Y"
                  "(4)\nCertify Receipt for Payment Max. Amount (HKD)": "Unlimited"
                  "Allow Further Delegation ": "Y"
                  "(I)\nEnquire BR - General (FMS)": "Y"
                  "(II)\nEnquire BR - Staffing related (HRMS)": "Y"
                  "(III)\nEnquire BR - Student related (SIS)": "Y"
                  "Allow Further Delegation": "Y"
                  "(A)\nApprove / Enquire Budget Commitment in ALL Systems": "Y"
                  "Approve Budget Commitment Max. Amount (HKD) in ALL Systems": "Unlimited"
                  "Valid From": "2025-01-07"
                  "Valid To": "31/12/2099"
              cads_canonical:
                summary: Canonical CADS payload
                value:
                  meta:
                    eventId: "CADS-user_a_ust.hk-16500-2025-01-07"
                    eventTime: "2025-01-07T00:00:00Z"
                    sourceSystem: "CADS"
                    correlationId: ""
                    idempotencyKey: "CADS|user_a@ust.hk|16500|BCO|2025-01-07"
                    operation: "ASSIGN_ENTITLEMENT"
                  identity:
                    email: "user_a@ust.hk"
                  entitlement:
                    roleName: "BCO"
                    departmentOrProject: "16500"
                    application: "SIS"
                    validFrom: "2025-01-07"
                    validTo: "2099-12-31"
                  attributes:
                    permissions:
                      enquireReqPoReceipt: true
                      recordReceiptGoodsServices: true
                      certifyReceiptForPayment: true
                      enquireBrGeneralFms: true
                    limits:
                      certifyReceiptForPaymentMaxAmountHkd: null
      responses:
        "202":
          $ref: '#/components/responses/Accepted'
        "422":
          description: CADS row failed transformation validation
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:  { type: string }
                  errors:
                    type: array
                    items: { type: string }
  /api/v1/inbound/cads/transform:
    post:
      summary: Dry-run CADS row transformation
      description: |
        Transform a raw CADS table row into canonical MidPoint JSON **without persisting** it.
        Use this endpoint to preview and validate the mapping before ingestion.

        Returns `200` with `{ isValid, errors, payload }` when the row is valid.
        Returns `422` when required fields are missing, with the partial payload included for debugging.
      operationId: transformCadsRow
      tags: [Ingestion]
      parameters:
        - $ref: '#/components/parameters/ApiKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw CADS table row with column headers as keys
              additionalProperties: true
            examples:
              cads_raw_row:
                summary: Raw CADS table row
                value:
                  "User Email": "user_a"
                  "Role": "BCO"
                  "Department / Project": "16500"
                  "(1)\nEnquire REQ/PO/\nReceipt ": "Y"
                  "(4)\nCertify Receipt for Payment Max. Amount (HKD)": "Unlimited"
                  "Allow Further Delegation ": "Y"
                  "(I)\nEnquire BR - General (FMS)": "Y"
                  "(III)\nEnquire BR - Student related (SIS)": "Y"
                  "Valid From": "2025-01-07"
                  "Valid To": "31/12/2099"
      responses:
        "200":
          description: Transformation succeeded — canonical payload returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  isValid:  { type: boolean, example: true }
                  errors:   { type: array, items: { type: string } }
                  payload:
                    $ref: '#/components/schemas/CadsCanonicalPayload'
        "422":
          description: Row failed required-field validation
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:    { type: string }
                  isValid:  { type: boolean, example: false }
                  errors:   { type: array, items: { type: string } }
                  payload:
                    $ref: '#/components/schemas/CadsCanonicalPayload'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/v1/inbound/peoplesoft:
    post:
      summary: Ingest PeopleSoft event (alias)
      description: "Alias for `POST /api/v1/inbound/events` with `X-Source-System: PEOPLESOFT`"
      operationId: ingestPeopleSoft
      tags: [Ingestion]
      parameters:
        - $ref: '#/components/parameters/ApiKey'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestPayload'
            examples:
              peoplesoft_row:
                summary: Raw PeopleSoft export row
                value:
                  "Dept": "ARO"
                  "Rank/ Team": "Prog Reg Team"
                  "User": "ar.registrar"
                  "Role Name": "HKUST STU SERVICE CENTER"
                  "Remarks": "Access to SIS"
                  "Data Level Security": "All students"
              peoplesoft_canonical:
                summary: Canonical PeopleSoft payload
                value:
                  meta:
                    sourceSystem: "PEOPLESOFT"
                    operation: "ASSIGN_ENTITLEMENT"
                  identity:
                    email: "ar.registrar@ust.hk"
                    userType: "staff"
                  entitlement:
                    application: "SIS"
                    action: "provision"
                    roleName: "HKUST STU SERVICE CENTER"
                    departmentOrProject: "ARO"
                  attributes:
                    rankOrTeam: "Prog Reg Team"
                    remarks: "Access to SIS"
                    dataLevelSecurity:
                      scope: "ALL_STUDENTS"
                      label: "All students"
      responses:
        "202":
          $ref: '#/components/responses/Accepted'
  /api/v1/inbound/ecm:
    post:
      summary: Ingest ECM event (alias)
      description: "Alias for `POST /api/v1/inbound/events` with `X-Source-System: ECM`"
      operationId: ingestECM
      tags: [Ingestion]
      parameters:
        - $ref: '#/components/parameters/ApiKey'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestPayload'
      responses:
        "202":
          $ref: '#/components/responses/Accepted'
...
  /api/v1/inbound/jspm:
    post:
      summary: Ingest JSPM event (alias)
      description: "Alias for `POST /api/v1/inbound/events` with `X-Source-System: JSPM`"
      operationId: ingestJSPM
      tags: [Ingestion]
      parameters:
        - $ref: '#/components/parameters/ApiKey'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestPayload'
      responses:
        "202":
          $ref: '#/components/responses/Accepted'

  # ── User endpoints ──────────────────────────────────────────────────────────
...
components:
  parameters:
    ApiKey:
      name: api_key
      in: header
      required: true
      description: API key for authentication
      schema:
        type: string
        example: poc-dev-key-1234
    XSourceSystem:
      name: X-Source-System
      in: header
      required: false
      schema:
        type: string
        enum: [CADS, PEOPLESOFT, ECM, JSPM]
    XCorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Unique key to prevent duplicate processing
      schema:
        type: string
...
  schemas:
    IngestPayload:
      type: object
      description: |
        Flexible payload — field names vary by source system.
        The platform normalises each source into a canonical format.
      additionalProperties: true
      example:
        employeeEmail: "john.doe@ust.hk"
        employeeId: "E12345"
        action: "provision"
    IngestResponse:
      type: object
      properties:
        message:       { type: string }
        eventId:       { type: string }
        jobId:         { type: string }
        status:        { type: string }
        correlationId: { type: string }
...
    EventStatus:
      type: object
      properties:
        eventId:       { type: string }
        jobId:         { type: string }
        sourceSystem:  { type: string }
        status:
          type: string
          enum: [received, validated, validation_failed, sent_to_midpoint, success, failed, retrying, dead_letter]
        correlationId: { type: string }
        identity:
          type: object
          properties:
            email:       { type: string }
            emailDomain: { type: string }
            displayName: { type: string }
            userType:    { type: string }
        validationErrors:
          type: array
          items: { type: string }
        retryCount:    { type: integer }
        lastError:     { type: string }
        receivedAt:    { type: string, format: date-time }
        processedAt:   { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error:  { type: string }
        detail: { type: string }
    CadsCanonicalPayload:
      type: object
      description: Canonical MidPoint JSON produced by the CADS transformer
      properties:
        meta:
          type: object
          properties:
            eventId:        { type: string }
            eventTime:      { type: string, format: date-time }
            sourceSystem:   { type: string, example: CADS }
            correlationId:  { type: string }
            idempotencyKey: { type: string }
            operation:      { type: string, example: ASSIGN_ENTITLEMENT }
        identity:
          type: object
          properties:
            email: { type: string, format: email }
        entitlement:
          type: object
          properties:
            roleName:             { type: string, example: BCO }
            departmentOrProject:  { type: string, example: "16500" }
            application:          { type: string, example: SIS }
            validFrom:            { type: string, format: date, nullable: true }
            validTo:              { type: string, format: date, nullable: true }
        attributes:
          type: object
          properties:
            permissions:
              type: object
              description: Boolean permission flags mapped from CADS Y/N columns
              additionalProperties: { type: boolean }
            limits:
              type: object
              description: Amount limits; null indicates Unlimited
              additionalProperties:
                type: number
                nullable: true
            delegation:
              type: object
              properties:
                procurement:
                  type: object
                  properties:
                    allowFurtherDelegation: { type: boolean }
                budget:
                  type: object
                  properties:
                    allowFurtherDelegation: { type: boolean }
      example:
        meta:
          eventId: "CADS-user_a_ust.hk-16500-2025-01-07"
          eventTime: "2025-01-07T00:00:00.000Z"
          sourceSystem: "CADS"
          correlationId: ""
          idempotencyKey: "CADS|user_a@ust.hk|16500|BCO|2025-01-07"
          operation: "ASSIGN_ENTITLEMENT"
        identity:
          email: "user_a@ust.hk"
        entitlement:
          roleName: "BCO"
          departmentOrProject: "16500"
          application: "SIS"
          validFrom: "2025-01-07"
          validTo: "2099-12-31"
        attributes:
          permissions:
            enquireReqPoReceipt: true
            recordReceiptGoodsServices: true
            certifyReceiptForPayment: true
            enquireBrGeneralFms: true
            enquireBrStaffingHrms: true
            enquireBrStudentSis: true
            approveEnquireBudgetCommitmentAllSystems: true
          limits:
            certifyReceiptForPaymentMaxAmountHkd: null
            approveBudgetCommitmentAllSystemsMaxAmountHkd: null
          delegation:
            procurement:
              allowFurtherDelegation: true
            budget:
              allowFurtherDelegation: true
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid api_key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Accepted:
      description: Event accepted for processing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/IngestResponse'